Fixed various mismatches in CpuCastKernel
* Fixes various mismatches when converting FP32 to BF16 and
  BF16 to FP32

* Fixed segfault when trying logging=1 and trying to log BF16

* Resolves MLCE-979

Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/8830
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>

+++ utils/TypePrinter.h
@@ -480,7 +480,7 @@ inline ::std::ostream &operator<<(::std::ostream &os, const BoundingBoxTransform
 inline ::std::ostream &operator<<(::std::ostream &os, const bfloat16 &v)
 {
     std::stringstream str;
-    str << v;
+    str << static_cast<float>(v);
     os << str.str();
     return os;
 }
